feat: Lotto 기능 추가
Lotto끼리 일치하는 숫자 개수 반환
예외처리 추가
1. 6자리만 가능
2. 중복 없어야 함
3. 오름차순이어야 함
